backend/requirements.txt
fastapi
requests
APScheduler
opml
feedparser
python-dateutil
SQLAlchemy
yt-dlp
mysql-connector-python
python-dotenv
uvicorn
alembic